requirements.txt
Django==2.2.14
django-cors-headers==3.7.0
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
pillow==8.2.0
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4


